requirements.txt
fastapi
uvicorn
pydantic
numpy
scipy
openai
python-decouple
